pyproject.toml
[tool.poetry]
name = "sscream_gendiff"
version = "0.2.0"
description = ""
authors = ["Sergey Sakirko <ssakirko@mail.ru>"]

packages = [
    { include = "gendiff" }
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
flake8 = "^3.8.4"
coverage = "^5.3"
mypy = "0.782"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
